+++ test/controllers/reports/designs_controller_test.rb
@@ -42,63 +42,63 @@ class Reports::DesignsControllerTest < ActionController::TestCase
 
   test 'should get advanced report before sheet date as viewer' do
     login(@viewer)
-    get :advanced, params: { project_id: @project, id: @design, sheet_before: '10/18/2012' }
+    post :advanced_report, params: { project_id: @project, id: @design, sheet_before: '10/18/2012' }
     assert_not_nil assigns(:design)
     assert_response :success
   end
 
   test 'should get advanced report after sheet date as viewer' do
     login(@viewer)
-    get :advanced, params: { project_id: @project, id: @design, sheet_after: '10/01/2012' }
+    post :advanced_report, params: { project_id: @project, id: @design, sheet_after: '10/01/2012' }
     assert_not_nil assigns(:design)
     assert_response :success
   end
 
   test 'should get advanced report between sheet date as viewer' do
     login(@viewer)
-    get :advanced, params: { project_id: @project, id: @design, sheet_after: '10/01/2012', sheet_before: '10/18/2012' }
+    post :advanced_report, params: { project_id: @project, id: @design, sheet_after: '10/01/2012', sheet_before: '10/18/2012' }
     assert_not_nil assigns(:design)
     assert_response :success
   end
 
   test 'should get advanced report by week as viewer' do
     login(@viewer)
-    get :advanced, params: { project_id: @project, id: @design, by: 'week' }
+    post :advanced_report, params: { project_id: @project, id: @design, by: 'week' }
     assert_not_nil assigns(:design)
     assert_response :success
   end
 
   test 'should get advanced report by year as viewer' do
     login(@viewer)
-    get :advanced, params: { project_id: @project, id: @design, by: 'year' }
+    post :advanced_report, params: { project_id: @project, id: @design, by: 'year' }
     assert_not_nil assigns(:design)
     assert_response :success
   end
 
   test 'should get advanced report with row variable (dropdown) as viewer' do
     login(@viewer)
-    get :advanced, params: { project_id: @project, id: @design, f: [{ id: variables(:one).id, axis: 'row', missing: '1' }] }
+    post :advanced_report, params: { project_id: @project, id: @design, f: [{ id: variables(:one).id, axis: 'row', missing: '1' }] }
     assert_not_nil assigns(:design)
     assert_response :success
   end
 
   test 'should get advanced report with row variable (dropdown) and exclude missing as viewer' do
     login(@viewer)
-    get :advanced, params: { project_id: @project, id: @design, f: [{ id: variables(:one).id, axis: 'row', missing: '0' }] }
+    post :advanced_report, params: { project_id: @project, id: @design, f: [{ id: variables(:one).id, axis: 'row', missing: '0' }] }
     assert_not_nil assigns(:design)
     assert_response :success
   end
 
   test 'should get advanced report with column variable (dropdown) as viewer' do
     login(@viewer)
-    get :advanced, params: { project_id: @project, id: @design, f: [{ id: variables(:one).id, axis: 'col', missing: '1' }] }
+    post :advanced_report, params: { project_id: @project, id: @design, f: [{ id: variables(:one).id, axis: 'col', missing: '1' }] }
     assert_not_nil assigns(:design)
     assert_response :success
   end
 
   test 'should get advanced report string (row) by sheet date (col) as viewer' do
     login(@viewer)
-    get :advanced, params: {
+    post :advanced_report, params: {
       project_id: @project, id: @design,
       f: [
         { id: variables(:string).id, axis: 'row', missing: '1' },
@@ -111,7 +111,7 @@ class Reports::DesignsControllerTest < ActionController::TestCase
 
   test 'should get advanced report with column variable (date) as viewer' do
     login(@viewer)
-    get :advanced, params: {
+    post :advanced_report, params: {
       project_id: @project, id: @design,
       f: [{ id: variables(:date).id, axis: 'col', missing: '1' }]
     }
@@ -121,7 +121,7 @@ class Reports::DesignsControllerTest < ActionController::TestCase
 
   test 'should get advanced report with column variable (numeric) as viewer' do
     login(@viewer)
-    get :advanced, params: {
+    post :advanced_report, params: {
       project_id: @project, id: designs(:all_variable_types),
       f: [{ id: variables(:numeric).id, axis: 'col', missing: '1' }]
     }
@@ -131,7 +131,7 @@ class Reports::DesignsControllerTest < ActionController::TestCase
 
   test 'should get advanced report gender (row) by weight (column) as viewer' do
     login(@viewer)
-    get :advanced, params: {
+    post :advanced_report, params: {
       project_id: @project, id: designs(:weight_and_gender),
       f: [
         { id: variables(:gender).id, axis: 'row', missing: '0' },
@@ -144,7 +144,7 @@ class Reports::DesignsControllerTest < ActionController::TestCase
 
   test 'should get advanced report weight (row) by site (column) as viewer' do
     login(@viewer)
-    get :advanced, params: {
+    post :advanced_report, params: {
       project_id: @project, id: designs(:weight_and_gender),
       f: [
         { id: variables(:weight).id, axis: 'row', missing: '0' },
@@ -157,7 +157,7 @@ class Reports::DesignsControllerTest < ActionController::TestCase
 
   test 'should get advanced report site and gender (row) by weight (column) as viewer' do
     login(@viewer)
-    get :advanced, params: {
+    post :advanced_report, params: {
       project_id: @project, id: designs(:weight_and_gender),
       f: [
         { id: 'site', axis: 'row', missing: '0' },